services/mailbox/requirements.txt
Flask
Flask-SQLAlchemy
pycryptodome == 3.15.0
gunicorn == 20.1.0
psycopg2-binary == 2.9.3
